api/requirements.txt
fastapi==0.94.0
pandas
uvicorn
psycopg2-binary
boto3